Hi All,
Please find attached a series of documents and codes corresponding to the STM project. We believe that we have the diffusion part ready, after tests with some analytical solution and another case. (We still need to test the Fortran code with other values of Theta, though.) Please read "Tests_for_the_code.doc" first.

Please let us know if you need any clarification on this regard.

It would be very nice if we could set up a meeting soon.

Best.

Fabian and Kaveh


git-svn-id: http://dminfo:8686/svn/repository/models/trunk@1249 cdc4813b-4270-ec4f-936e-925f93a782c2

Former-commit-id: 982c08625a71de6e0ffc5176e5cac00b3a6abd66