-The results are not included diffusion (I made it comment) -The wiggles are due to excluding flux limiter (as far as I found these days in literature superbee is suitable for our case but I am going to try different types and check the one best fitted our case) -For a dense mesh results are as follows:
Numerical diffusion is almost vanished 
(I sent a sin in with max =1.00 and received at the end of 100 m channel sin wave with max= 0.99701)

Position of center of mass 
in 100 m channel it changed  18.8 cm ( error is 18/100000) and I think some of this error belongs to wiggles (a flux limiter may treat this error)

I will re write this in Fortran and send to you sometimes next week 

I have a question:
In the CFD forums and some literature I encountered cases which a constant number was applied instead of a function, as a flux limiter

Flux = Flux L -  Constant (Flux H -Flux L), and good results were  received, I want to know your Idea about this case?


Finally you may find the results and debugged code(thanks to Fabian again and again)  attached.

Have a nice weekend,
Kaveh 


git-svn-id: http://dminfo:8686/svn/repository/models/trunk@1414 cdc4813b-4270-ec4f-936e-925f93a782c2

Former-commit-id: 8efc6a6bbc8be90f8e49afebefd1cd56b7cc596d