First cut of DSM2-PTM2, which is a C++ version of DSM2-PTM.  A major reason for this task is to make maintenance easy for the future.   This version is written in C++ without relying on Fortran I/O code, so it makes debugging somewhat easy but format change in the tide file must be reflected here separately.
The functionality of this version for now is very identical to the current Java DSM2-PTM.  The input files are compatible.

Major changes are:
1. Simulation and postprocessing are separated. It means that a user can get new output items without running a same simulation with identical tides.  Also, the particle number limit that the Java version has is resolved.  A run with more than 100,000 particles was successful even though the size of an output file can be larger than a few gigabytes.
2. The raw outputs are consolidated into one HDF5 file.  This is better for postprocessing and maintenance of runs.

Somethings are not implemented yet:
1. Object-to-object is not implemented.
2. Behaviors are not implemented.
3. Animation is not supported yet.  Particle locations are recorded, but no animation tool is 
4. Lots of comments and tests need to be added.

The difference in results from PTM and PTM2 is not studied systematically yet.  Rough comparison reveals some differences, but overall trends in reveals are similar.  There are a few possible causes of the differences.  C++ and Java may handle numbers slightly differently, and the two versions use different random number generators.  Further tests and comparisons are absolutely necessary.

git-svn-id: http://dminfo:8686/svn/repository/models/trunk@2105 cdc4813b-4270-ec4f-936e-925f93a782c2

Former-commit-id: 82919eebd5aef681ec341771bbf412a76a7c5380