Resolve an issue in the Debug build
The debug build fails because of library conflicts. To resolve that
libmmdd.lid is excluded after the flags settings are seprated into DEBUG
and RELEASE. It still does not resolved the issue with PTM.dll due to
the conflict of powf in libcpmtd.lib and libmmt.lib in the DEBUG build.
I found no solution for this yet, so I will leave it at that.