Remove name conflicts from merging common_xsect
These conflicts occur because DSM2 and GTM each had a module named
'common_xsect'. Now that these modules are combined into one,
and because a name may be assigned to different things
(e.g., 'virt_xsect' is a module in DSM2 but a variable in GTM),
these conflicts need to be resolved on a case-by-case basis.