Update data type for 64-bit
This commit updates the data type of some variables in the klu interface
for 64-bit. Somehow keeping Fortran integer for those in 32-bit way
works fine for the standalone binaries, but it causes seg faults for
the Python so file. Adjusting those pointer variables to the 64-bit size
resolves the issue. It may need to be updated later to make it work for
both architectures.