Export fourpt for Python
This commit adds a module to export only parts that we want to call from
Python. It also changes the name of the export module and updates the
entry point and the driver of fourpt slightly. The names of modules and
export functions are tentative.
DSM2 hydro can be called from Python as below:
import pydsm2
pydsm2.pydsm2hydro.main('hydro.inp')