Rewrite CMake files (only for Windows for now)
The goals of this CMake file updates are: Cleaning up the CMake
files, using recent versions of CMake, accommodating different
platforms better.
This commit is written only for Windows 64-bit for now, and will not
work for Linux.
To the build process faster, this commit rewrites input_storage
CMakeLists.txt to use generated codes in the CMake style, which will
save generated codes into the build directory, instead of the
source directory. To do so, generate.py needed to be updated slightly
not to write into the source directory.