Consolidate CMake projects into one
Previously, we had three separate projects, oprule, input_storage, and
DSM2. This commit consolidates all into just one project. The DSM2
dependencies, oprule and input_storage, are not used for other projects,
and this makes the CMake simple and robust.
Also, a short documentation about how to build is added, and the CMake
version is bumped.

Closes DSM2-808