Remove JRE in CPack
We used to pack JRE together, but Java does not provide a JRE anymore.
So, now it is not included in a CPack file. The PTM batch and script
files are updated to find Java from JAVA_HOME.
The Java version requirement is removed in CMake because any version
of Java after 1.6 seems to work.

DSM2-854