Make Hydro and GTM side by side
This commit add a new project hydro_gtm to couple Hydro and GTM. It
will create a new binary hydro_gtm.
The new binary does not usefully couple the two modules yet. It only
runs one step of Hydro and one step of GTM in a row. None of variables
are shared in memory, except some of the grid information. Data are
passed through the tidefile, just like GTM is running uncoupled. No GTM
data can be fed back to Hydro yet. So, this is just the first step for a
real coupling.
For coupling, the frequency of tidefile reading in GTM is adjusted to
read hydro outputs from the tidefile every time step, not a bunch of
steps.

DSM2-684