Package Intel and some other runtime
This commits tries to cpack Intel and some other runtime libraries on
Linux. On Linux, Boost cannot be linked statically in some cases, and this
should reduce the inconvenience with the runtime libraries. No changes
in Windows build.
The location of PTM artifacts are moved to lib directory.
CMake files are cleaned and reformatted slightly.

DSM2-874