Add SIZE_T for HDF5
It turns out that SIZE_T from HDF5 can be 8-byte, which can be different
from Fortran integer. This mismatch can lead to a HDF5 issue. In the
previous commit, some SIZE_T are introduced to remove type warnings, but
they were not complete. This commit fixes this issue.
It is likely to be more of these data type issues lurking somewhere, and
they will be addressed later.