Introduce stdlib loggers
This commit introduces stdlib loggers to use a better logging and to
replace the current logging. Two loggers, `logger` and `stderr_logger`,
are added to `logging` module. The `logger` writes to a file, `dsm2.log`
and to stdout, and the `stderr_logger` writes to the same file and the
stderr.