Reorganize codes - only for Qual
This commit is rather large as the first effort to reorganize the DSM2
code base. The current DSM2 code is the accumulation of changes started
from FORTRAN 77 code, which uses outdated styles and global variables
and functions. We want to update the code to help future improvement by
introducing modern Fortran and removing intertwined parts.
This commit
finished reorganization only for Qual. No functional changes are
introduced. The reorg ended up into a large change because one change
led to another to make it work.  The codes are modulized and moved
aroundi while the original formatting is kepet. Some variables and
procedures names changed to avoid name conflicts. Some comments are
added to help refactoring.