Modularize Hydro
The Hydro code is modularized to organize the code better with less name
confusions and conflicts. Modularization of the Fortran codes led to
changes in the C interface names. Instead of updating the interface
names with the module prefixes, modern standard ISO C bind are
introduced.
Also, Hydro driver is separated.
`chan_ec_val` routines are removed temporarily to move out to hydro_gtm
later.
No functional changes are introduced.